**2.0.1-Build 12 & 13** * Fixed an issue causing all following loaded shops to not display if a shop is within an unloaded chunk or invalid world * Fixed potential issue causing shop items to not return to the player when destroyed * Fixes for empty inventory titles from configuration * SetTitle was set as async for updating display entities

+++ Core/src/main/java/xzot1k/plugins/ds/api/objects/DShop.java
@@ -19,6 +19,7 @@ import xzot1k.plugins.ds.api.enums.EconomyCallType;
 import xzot1k.plugins.ds.api.events.EconomyCallEvent;
 import xzot1k.plugins.ds.api.events.ShopVisitEvent;
 import xzot1k.plugins.ds.api.handlers.DisplayPacket;
+import xzot1k.plugins.ds.core.packets.Display;
 
 import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
@@ -512,8 +513,10 @@ public class DShop implements Shop {
      */
     public synchronized void delete(boolean async) {
 
-        Display display = DisplayShops.getPluginInstance().getDisplayManager().getDisplay(getShopId());
-        if (display != null) {display.delete();}
+        if (DisplayShops.getPluginInstance().getDisplayManager() != null) {
+            Display display = DisplayShops.getPluginInstance().getDisplayManager().getDisplay(getShopId());
+            if (display != null) {display.delete();}
+        }
 
         if (async) INSTANCE.getServer().getScheduler().runTaskAsynchronously(INSTANCE, (Runnable) this::delete);
         else delete();
